+++ cloud/blockstore/libs/storage/partition/model/mixed_blocks_filter_ut.cpp
@@ -1,4 +1,4 @@
-#include "mixed_index_blocks_filter.h"
+#include "mixed_blocks_filter.h"
 
 #include <library/cpp/testing/unittest/registar.h>
