+++ cloud/blockstore/libs/storage/partition/model/ut/ya.make
@@ -15,7 +15,7 @@ SRCS(
     fresh_blob_ut.cpp
     garbage_queue_ut.cpp
     group_downtimes_ut.cpp
-    mixed_index_blocks_filter_ut.cpp
+    mixed_blocks_filter_ut.cpp
     mixed_index_cache_ut.cpp
 )
