+++ cloud/blockstore/libs/storage/partition/part_cleanup_logic_ut.cpp
@@ -83,7 +83,7 @@ TPartitionState MakeState(size_t blockCount = 2048)
         1,             // compactionRangeCountPerRun
         std::move(threadSafeState),
         TTestExecutor::TabletId,
-        false);        // mixedIndexBlocksFilterEnabled
+        false);        // mixedBlocksFilterEnabled
 }
 
 NProto::TBlobMeta MakeMixedBlobMeta(
