+++ cloud/blockstore/libs/storage/partition/part_state_ut.cpp
@@ -119,7 +119,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         const auto initialBackpressure = state.CalculateCurrentBackpressure();
@@ -191,7 +191,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         state.GetCompactionMap().Update(0, 30, 30, 30, 0, false);
@@ -229,7 +229,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         state.GetLogicalUsedBlocks().Set(0, 9);
@@ -317,7 +317,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         state.IncrementMergedBlocksCount(5_GB / DefaultBlockSize);
@@ -363,7 +363,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         TTestExecutor executor;
@@ -505,7 +505,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,   // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
         UNIT_ASSERT_VALUES_EQUAL(maxBlobsPerDisk, state.GetMaxBlobsPerDisk());
     }
@@ -541,7 +541,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         TCleanupQueueItem b1 {{1, 1, 4, 4_MB, 0, 0}, 111, {}};
@@ -589,7 +589,7 @@ Y_UNIT_TEST_SUITE(TPartitionStateTest)
             1,      // compactionRangeCountPerRun
             threadSafeState,
             0,      // tabletId
-            false   // mixedIndexBlocksFilterEnabled
+            false   // mixedBlocksFilterEnabled
         );
 
         const ui32 blockIndex = 0;
